Added structure to parser, separated parselets to own files
