+++ src/parser/p_d_parens.rs
@@ -1,0 +1,38 @@
+use crate::parser::{Parselet, ParseError, parse_expression};
+use crate::lexer::Lexer;
+use crate::ast::Expression;
+
+pub struct LeftParensParselet {}
+
+impl Parselet for LeftParensParselet {
+    fn parse(&self, lexer: &mut Lexer) -> Result<Box<dyn Expression>, ParseError> {
+        parse_expression(0, lexer)
+    }
+
+    fn nud(&self, lexer: &mut Lexer) -> Result<Option<Box<dyn Expression>>, ParseError> {
+        let expr = parse_expression(
+            0,
+            lexer)?;
+        Ok(Some(expr))
+    }
+
+    fn led(&self, _lexer: &mut Lexer, _left: Box<dyn Expression>) -> Result<Option<Box<dyn Expression>>, ParseError> {
+        Err( ParseError { msg: "Can't parse ( in NUD position".to_string() } )
+    }
+}
+
+pub struct RightParensParselet {}
+
+impl Parselet for RightParensParselet {
+    fn parse(&self, lexer: &mut Lexer) -> Result<Box<dyn Expression>, ParseError> {
+        parse_expression(0, lexer)
+    }
+
+    fn nud(&self, _lexer: &mut Lexer) -> Result<Option<Box<dyn Expression>>, ParseError> {
+        Err( ParseError { msg: "Can't parse ) in prefix position".to_string() } )
+    }
+
+    fn led(&self, _lexer: &mut Lexer, left: Box<dyn Expression>) -> Result<Option<Box<dyn Expression>>, ParseError> {
+        Ok(Some(left))
+    }
+}
